+++ busTrackerApi/src/main/kotlin/busTrackerApi/db/models/Shape.kt
@@ -4,5 +4,6 @@ data class Shape(
     val itineraryId: String,
     val latitude: Double,
     val longitude: Double,
-    val sequence: Int
+    val sequence: Int,
+    val distance: Double
 )